Updated for fall 2020
Hey, I came across your tool and found it pretty useful. I had to change the list of past midterms as well as the link to work for fall 2020. But it worked for me despite being 2 years old haha. So I'm proposing those changes here in hopes it can help more students.